requirements.txt
aiogram==3.15.0
openai
supabase==2.11.0
python-dotenv==1.0.1
apscheduler==3.10.4
